fix(draugr/arena): widen tuner_adapt accumulators to uint64_t
total_small and total_large accumulate up to 8×2³² and 32×2³²
respectively. Multiplying by 100 overflows uint32_t (~4.3×10⁹) when
total_small exceeds ~42M samples, producing wrong ratios and causing
the tuner to set incorrect frequency thresholds.

- Widen total_small, total_large, and total from uint32_t to uint64_t

BUG #2 (wbuf_flush ring-full data loss): not a bug — entries before
the full check are written to the ring; remaining entries are lost
because the ring is full, which is the intended design (seal wbuf).

BUG #3 (bare_spill_find_all iterator invalidation): not triggerable
— bare_spill_find_all takes const ht_bare_t*, so callbacks cannot
mutate the table through that pointer. No callback in the codebase
mutates during iteration.

BUG #4 (ht_cache_remove + concurrent evict): not a code bug —
ht_cache_t is documented as NOT thread-safe; external synchronization
required.

ISSUE #5 (iter_next returns evicted entry): not a code bug — same
thread-safety contract applies to iterators.

+++ src/arena.c
@@ -701,11 +701,11 @@ static void tuner_adapt(struct arena *a) {
     struct arena_tuner *t = &a->tuner;
     if (t->sample_count < 1024) return;
 
-    uint32_t total_small = 0, total_large = 0;
+    uint64_t total_small = 0, total_large = 0;
     for (int i = 0; i < 8; i++) total_small += t->hist[i];
     for (int i = 32; i < 64; i++) total_large += t->hist[i];
 
-    uint32_t total = t->sample_count;
+    uint64_t total = t->sample_count;
     if (total == 0) return;
 
     if (total_small * 100 / total > 70) {
